backend/requirements.txt
sqlalchemy==2.0.25
fastapi==0.110.0
psycopg2-binary~=2.8.6
uvicorn==0.27.0
python-dotenv
bcrypt==4.2.0
passlib==1.7.4
python-jose==3.3.0
PyJWT==2.8.0
# Add more here
